Clean Indexation's HashedIndex
